direnv: make sure the right go tools are in the path
Even though we were setting GOROOT to an explicit version of go, this
was not added to our path. So if multiple versions of go are installed
your global version will be used.

Add GOROOT to the front of the path to ensure that when working with the
project we are using the right tooling.

Signed-off-by: Gavin-John Noonan <mail@gjnoonan.co.uk>