improve error handling for switch delete
there's a lot here, let's break it down.

this uses the richer error model for gRPC described here:

https://grpc.io/docs/guides/error/#richer-error-model

and here:

https://google.aip.dev/193#error_model

and using examples from:

https://github.com/grpc/grpc-go/tree/master/examples/features/error_details

There's a lot more of this kind of thing to be done in the gRPC api.
This is just a first example.

I updated golangci-lint and had to cope with the fallout from that,
including a lot of gosec related cast issues.Fixing those also lead to
forcing all TCP port values to be uint16 as it seems they should be,
keeping disk/iso file sizes as uint64 and some other cleanups.

Having the newer golangci-lint meant the ability to add
google.golang.org/grpc/internal/status to ignorePackageGlobs which was
needed for the gRPC error handling changes.